déplacement de toutes les méthodes dans des classes spécifiques
