+++ src/test/java/org/example/volunteers/EmailTest.java
@@ -11,28 +11,28 @@ public class EmailTest {
     @Test
     public void validEmailAddressShouldVeValidated() {
         String emailAddress = "adresse.mail@mail.com";
-        boolean isEmail = Cleaner.isValidEmail(emailAddress);
+        boolean isEmail = Email.isValidEmail(emailAddress);
         assertTrue(isEmail, "Une adresse email valide est validée");
     }
 
     @Test
     public void invalidEmailShouldNotBeValidated() {
         String emailAddress = "adressemailmailcom";
-        boolean isEmail = Cleaner.isValidEmail(emailAddress);
+        boolean isEmail = Email.isValidEmail(emailAddress);
         assertFalse(isEmail, "Une adresse email invalide n'est pas validée");
     }
 
     @Test
     public void addressCleanerShouldReplaceCorrectCharacters() {
         String toClean = "éèëêœàäâîïìûùüôòö";
-        toClean = Cleaner.cleanEmailAddress(toClean);
+        toClean = Email.cleanEmailAddress(toClean);
         assertEquals("eeeeoeaaaiiiuuuooo", toClean, "L'addresse email est nettoyée des caractères avec accents");
     }
 
     @Test
     public void addressCleanerShouldNotReplaceNormalCharacters() {
         String toClean = "eeeeoeaaaiiiuuuooo";
-        toClean = Cleaner.cleanEmailAddress(toClean);
+        toClean = Email.cleanEmailAddress(toClean);
         assertEquals("eeeeoeaaaiiiuuuooo", toClean, "L'addresse email n'est pas affectée si elle contient des caractères normaux");
     }
 
@@ -40,7 +40,7 @@ public class EmailTest {
     public void mailAddressCleanupShouldSetEmptyAddressIfInvalid() {
         List<Volunteer> volunteers = new ArrayList<>();
         volunteers.add(new Volunteer("Nom", "Prenom", "pseudo", "email@gmailcom", "+33600000000"));
-        volunteers = Cleaner.cleanupMailAddresses(volunteers);
+        volunteers = Email.cleanupMailAddresses(volunteers);
         assertEquals("", volunteers.get(0).eMail, "L'adresse email invalide se retrouve vide");
     }
 
@@ -48,7 +48,7 @@ public class EmailTest {
     public void mailAddressCleanupShouldLowerCaseAddress() {
         List<Volunteer> volunteers = new ArrayList<>();
         volunteers.add(new Volunteer("Nom", "Prenom", "pseudo", "eMail@GMail.com", "+33600000000"));
-        volunteers = Cleaner.cleanupMailAddresses(volunteers);
+        volunteers = Email.cleanupMailAddresses(volunteers);
         assertEquals("email@gmail.com", volunteers.get(0).eMail, "L'adresse email invalide se retrouve vide");
     }
 
@@ -56,7 +56,7 @@ public class EmailTest {
     public void mailAddressCleanupShouldReplaceAccentedCharacters() {
         List<Volunteer> volunteers = new ArrayList<>();
         volunteers.add(new Volunteer("Nom", "Prenom", "pseudo", "dédé-la-bagarre@gmail.com", "+33600000000"));
-        volunteers = Cleaner.cleanupMailAddresses(volunteers);
+        volunteers = Email.cleanupMailAddresses(volunteers);
         assertEquals("dede-la-bagarre@gmail.com", volunteers.get(0).eMail, "L'adresse email contenant des caractères spéciaux est remplacée");
     }
 }
